refactor(cli): main.py 모듈 분리 (Phase 0 — 기능 변경 없음)
4937줄의 main.py를 책임별 서브모듈로 분리하여 557줄로 슬림화.
외부 import 표면(tubearchive.app.cli.main.*)은 re-export로 완전히 보존.

변경 내용:
- parser.py: create_parser, parse_schedule_datetime 이동
- validators.py: ValidatedArgs, validate_args, _resolve_* 이동
- pipeline.py: run_pipeline, TranscodeOptions/Result, 파이프라인 헬퍼 전체 이동
- upload.py (신규): upload_to_youtube, _upload_after_pipeline, cmd_upload_only 등 이동
- youtube.py: cmd_setup_youtube, cmd_youtube_auth, cmd_list_playlists 이동
- watch.py: _run_watch_mode, _setup_file_observer 등 이동
- status.py (신규): cmd_status, cmd_reset_build, cmd_fix_device_models 등 이동
- test_cli.py: watch 관련 patch 경로 watch.* 로 갱신
- tui/bridge.py, tui/screens/pipeline.py: 직접 소스 모듈 import로 전환

mock 표면 보존:
- init_database / database_session / MergeJobRepository — main.py 잔류
- 분리 모듈은 main 경유 lazy import로 patch 전파 유지
